0.10.4

+++ lib/src/gridfs/gridfs.dart
@@ -1,7 +1,7 @@
 part of '../../mongo_dart.dart';
 
 class GridFS {
-  static int defaultChunkSize = 256 * 1024;
+  static Int32 defaultChunkSize = Int32(256 * 1024);
   static int maxChunkSize = (3.5 * 1000 * 1000).toInt();
 
   Db database;
